refactor: Update import paths for AI related actions in chat components
